123 Infographic support to show learning event pages
Learning events are different types of objects. I modified the
learning event component to send the event to a manager class to
then call the appropriate component depending on the type of
event. In this case I was building functionality for Page type,
which would match for infographics stored in canvas.