SBAELIDEV-122 Created Component to handle video
Created a new route in the LearningEventsManager to redirect to the
LearningEventsFile component if the LearningEvent is a file. In the
LearningEventsFile component there's a control statement to redirect to
LearningEventsVideo component if the file is a video. This is the basis
of getting MLE videos up on the front-end.