Added Queryable concern
The queryable concern can be added to a regular ruby class. It adds
methods for doing find_by_sql, paginate_by_sql and count_by_sql.
These method execute a string of SQL, then populate an object with
the data. This concern is for use in cases where the result set of
a query may not match up with a model/table.